refactor: move reviewer check to FundingRoundService
